Initial Commit
This is (more or less) the program which brought us to the completed rebuild of
NAPDB.  It does everything we currently need it to do in terms of database
administration albeit somewhat rough round the edges.  It also contains a lot of
code only relevant to the rebuild which is now obsolete and will be removed as
we move towards a v1.0 release. [2002]