Better Error Catching on Data Import
If an import csv file is missing then, to guard against data loss, the import
process will also skip the table restructure as well as the actual data import.
In case of such an error, the program will move on to the next table to be
processed (if any) rather than terminating altogether.

Also in this update: remove the ‘districts’ transition process. [2010]