Add Safeguard to Local Govt Reorg Process
Unless overridden by options (-q or -b) the user will be prompted to confirm the
irrevocable changes she is about to make to the database.  She can choose to
barrel on regardless, perform a database backup first, revert to dry-run mode or
abort altogether.  The -b option ensures a backup can be requested even in quiet
mode. [1373]