Sui Generis Index Names
There are some entity names that we may wish to index in unique ways outwith the
base rules and any special index rules we may apply more generally.  The
EntityName class deals with this by taking, as a second parameter, a string
representing any existing index name which is preserved through the indexation
process if (one of) the special index rule(s) is the Sui Generis rule.  As an
example of circumstances in which we cannot apply a standard rule, we have the
county of West Yorkshire which we wish to index as yorkshirewest but the county
of West Midlands, which we wish to index as westmidlands. [1528]